Refactor and rename the Main window
