+++ docoloco/window/__init__.py
@@ -1,3 +1,3 @@
-from .application import ApplicationWindow
+from .main import MainWindow
 from .doc_page import DocPage
 from .new_page import NewPage
